[Xiaomi][Warning] drivers: sdm845-gpu: Buck based on silent frequency sdm845
Inspired by @xzr467706992's sm8250 GPU overclocking and the buck list shared on the website.

Unfortunately, the voltmeter of the SDM845 platform is hard-coded in the kernel, so it can only be implemented in a dirty way.

The working frequencies of most GPUs in mobile phones are around 180Mhz and 257Mhz, and high-load UI resources will be upscaled to 414Mhz, such as cameras.
So only for 257Mhz-342Mhz-414Mhz frequency to step down
257Mhz VDD_GX_LOWER ==> VDD_GX_MIN
342Mhz VDD_GX_LOW ==> VDD_GX_LOWER
414Mhz VDD_GX_LOW_L1 ==> VDD_GX_LOW

Signed-off-by: stic-server-open <1138705738@qq.com>